backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
python-dotenv
passlib[bcrypt]
flask-jwt-extended
requests
